+++ retz-client/src/main/java/io/github/retz/cli/CommandGetFile.java
@@ -73,7 +73,6 @@ public class CommandGetFile implements SubCommand {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
-                .enableAuthentication(fileConfig.authenticationEnabled())
                 .setAuthenticator(fileConfig.getAuthenticator())
                 .checkCert(fileConfig.checkCert())
                 .build();