+++ retz-client/src/main/java/io/github/retz/cli/CommandKill.java
@@ -48,7 +48,6 @@ public class CommandKill implements SubCommand {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
-                .enableAuthentication(fileConfig.authenticationEnabled())
                 .setAuthenticator(fileConfig.getAuthenticator())
                 .checkCert(fileConfig.checkCert())
                 .build()) {