+++ retz-client/src/main/java/io/github/retz/cli/CommandList.java
@@ -45,7 +45,6 @@ public class CommandList implements SubCommand {
         LOG.info("Configuration: {}", fileConfig.toString());
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
-                .enableAuthentication(fileConfig.authenticationEnabled())
                 .setAuthenticator(fileConfig.getAuthenticator())
                 .checkCert(fileConfig.checkCert())
                 .build()) {