+++ retz-client/src/main/java/io/github/retz/cli/CommandLoadApp.java
@@ -115,7 +115,6 @@ public class CommandLoadApp implements SubCommand {
                 enabled);
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
-                .enableAuthentication(fileConfig.authenticationEnabled())
                 .setAuthenticator(fileConfig.getAuthenticator())
                 .checkCert(fileConfig.checkCert())
                 .build()) {